+++ src/main/kotlin/snyk/common/annotator/SnykAnnotator.kt
@@ -1,32 +1,49 @@
-package snyk.code.annotator
+package snyk.common.annotator
 
 import com.intellij.codeInsight.intention.IntentionAction
 import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.ExternalAnnotator
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.Disposable
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.application.invokeLater
 import com.intellij.openapi.diagnostic.logger
+import com.intellij.openapi.editor.colors.TextAttributesKey
+import com.intellij.openapi.editor.markup.GutterIconRenderer
+import com.intellij.openapi.project.DumbAware
+import com.intellij.openapi.project.guessProjectForFile
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiFile
+import icons.SnykIcons
+import io.snyk.plugin.Severity
 import io.snyk.plugin.getSnykCachedResultsForProduct
+import io.snyk.plugin.getSnykToolWindowPanel
 import io.snyk.plugin.toLanguageServerURL
 import org.eclipse.lsp4j.CodeActionContext
 import org.eclipse.lsp4j.CodeActionParams
 import org.eclipse.lsp4j.Range
 import org.eclipse.lsp4j.TextDocumentIdentifier
-import snyk.code.annotator.SnykAnnotator.SnykAnnotation
 import snyk.common.AnnotatorCommon
 import snyk.common.ProductType
+import snyk.common.annotator.SnykAnnotationAttributeKey.critical
+import snyk.common.annotator.SnykAnnotationAttributeKey.high
+import snyk.common.annotator.SnykAnnotationAttributeKey.low
+import snyk.common.annotator.SnykAnnotationAttributeKey.medium
+import snyk.common.annotator.SnykAnnotationAttributeKey.unknown
+import snyk.common.annotator.SnykAnnotator.SnykAnnotation
 import snyk.common.lsp.LanguageServerWrapper
 import snyk.common.lsp.ScanIssue
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.TimeoutException
+import javax.swing.Icon
 
 private const val CODEACTION_TIMEOUT = 5000L
 
 abstract class SnykAnnotator(private val product: ProductType) :
-    ExternalAnnotator<Pair<PsiFile, List<ScanIssue>>, List<SnykAnnotation>>(), Disposable {
+    ExternalAnnotator<Pair<PsiFile, List<ScanIssue>>, List<SnykAnnotation>>(), Disposable, DumbAware {
+
     val logger = logger<SnykAnnotator>()
     protected var disposed = false
         get() {
@@ -40,10 +57,12 @@ abstract class SnykAnnotator(private val product: ProductType) :
     }
 
     inner class SnykAnnotation(
+        val issue: ScanIssue,
         val annotationSeverity: HighlightSeverity,
         val annotationMessage: String,
         val range: TextRange,
-        val intention: IntentionAction
+        val intention: IntentionAction,
+        val renderGutterIcon: Boolean = false
     )
 
     // overrides needed for the Annotator to invoke apply(). We don't do anything here
@@ -60,7 +79,8 @@ abstract class SnykAnnotator(private val product: ProductType) :
         if (!LanguageServerWrapper.getInstance().isInitialized) return emptyList()
 
         val annotations = mutableListOf<SnykAnnotation>()
-        initial.second.forEach { issue ->
+        val gutterIcons: MutableSet<TextRange> = mutableSetOf()
+        initial.second.sortedByDescending { it.getSeverityAsEnum() }.forEach { issue ->
             val textRange = textRange(initial.first, issue.range)
             val highlightSeverity = issue.getSeverityAsEnum().getHighlightSeverity()
             val annotationMessage = issue.annotationMessage()
@@ -69,14 +89,17 @@ abstract class SnykAnnotator(private val product: ProductType) :
                 return@forEach
             }
             if (!textRange.isEmpty) {
-                annotations.add(
-                    SnykAnnotation(
-                        highlightSeverity,
-                        annotationMessage,
-                        textRange,
-                        ShowDetailsIntentionAction(annotationMessage, issue)
-                    )
+                val detailAnnotation = SnykAnnotation(
+                    issue,
+                    highlightSeverity,
+                    annotationMessage,
+                    textRange,
+                    ShowDetailsIntentionAction(annotationMessage, issue),
+                    renderGutterIcon = !gutterIcons.contains(textRange)
                 )
+                annotations.add(detailAnnotation)
+                gutterIcons.add(textRange)
+
                 val params =
                     CodeActionParams(
                         TextDocumentIdentifier(initial.first.virtualFile.toLanguageServerURL()),
@@ -102,14 +125,14 @@ abstract class SnykAnnotator(private val product: ProductType) :
                     .sortedBy { it.right.title }.forEach { action ->
                         val codeAction = action.right
                         val title = codeAction.title
-                        annotations.add(
-                            SnykAnnotation(
-                                highlightSeverity,
-                                title,
-                                textRange,
-                                CodeActionIntention(issue, codeAction, product)
-                            )
+                        val codeActionAnnotation = SnykAnnotation(
+                            issue,
+                            highlightSeverity,
+                            title,
+                            textRange,
+                            CodeActionIntention(issue, codeAction, product)
                         )
+                        annotations.add(codeActionAnnotation)
                     }
 
             }
@@ -124,13 +147,28 @@ abstract class SnykAnnotator(private val product: ProductType) :
     ) {
         if (disposed) return
         if (!LanguageServerWrapper.getInstance().isInitialized) return
-        annotationResult.forEach { annotation ->
-            if (!annotation.range.isEmpty) {
-                holder.newAnnotation(annotation.annotationSeverity, annotation.annotationMessage)
-                    .range(annotation.range)
-                    .withFix(annotation.intention)
-                    .create()
+        annotationResult.sortedByDescending { it.issue.getSeverityAsEnum() }
+            .forEach { annotation ->
+                if (!annotation.range.isEmpty) {
+                    val annoBuilder = holder.newAnnotation(annotation.annotationSeverity, annotation.annotationMessage)
+                        .range(annotation.range)
+                        .textAttributes(getTextAttributeKeyBySeverity(annotation.issue.getSeverityAsEnum()))
+                        .withFix(annotation.intention)
+                    if (annotation.renderGutterIcon) {
+                        annoBuilder.gutterIconRenderer(SnykShowDetailsGutterRenderer(annotation))
+                    }
+                    annoBuilder.create()
+                }
             }
+    }
+
+    private fun getTextAttributeKeyBySeverity(severity: Severity): TextAttributesKey {
+        return when(severity) {
+            Severity.UNKNOWN -> unknown
+            Severity.LOW -> low
+            Severity.MEDIUM -> medium
+            Severity.HIGH -> high
+            Severity.CRITICAL -> critical
         }
     }
 
@@ -178,4 +216,54 @@ abstract class SnykAnnotator(private val product: ProductType) :
             return TextRange.EMPTY_RANGE
         }
     }
+}
+
+class SnykShowDetailsGutterRenderer(val annotation: SnykAnnotation) : GutterIconRenderer() {
+    override fun equals(other: Any?): Boolean {
+        return annotation == other
+    }
+
+    override fun hashCode(): Int {
+        return annotation.hashCode()
+    }
+
+    override fun getIcon(): Icon {
+        return SnykIcons.getSeverityIcon(annotation.issue.getSeverityAsEnum())
+    }
+
+    override fun getClickAction(): AnAction? {
+        if (annotation.intention !is ShowDetailsIntentionAction) return null
+        return getShowDetailsNavigationAction(annotation.intention)
+    }
+
+    private fun getShowDetailsNavigationAction(intention: ShowDetailsIntentionAction) =
+        object : AnAction() {
+            override fun actionPerformed(e: AnActionEvent) {
+                invokeLater {
+                    val virtualFile = intention.issue.virtualFile ?: return@invokeLater
+                    val project = guessProjectForFile(virtualFile) ?: return@invokeLater
+                    val toolWindowPanel = getSnykToolWindowPanel(project) ?: return@invokeLater
+
+                    intention.selectNodeAndDisplayDescription(toolWindowPanel)
+                }
+            }
+        }
+
+    override fun getTooltipText(): String {
+        return annotation.annotationMessage
+    }
+
+    override fun getAccessibleName(): String {
+        return annotation.annotationMessage
+    }
+
+    override fun isNavigateAction(): Boolean {
+        return true
+    }
+
+    override fun isDumbAware(): Boolean {
+        return true
+    }
+
+
 }