+++ src/main/kotlin/snyk/common/lsp/LanguageServerWrapper.kt
@@ -141,7 +141,7 @@ class LanguageServerWrapper(
                 if (!disposed) {
                     try {
                         process.errorStream.bufferedReader().forEachLine { println(it) }
-                    } catch (ignored: RuntimeException) {
+                    } catch (ignored: Exception) {
                         // ignore
                     }
                 }