+++ src/main/kotlin/snyk/common/lsp/Types.kt
@@ -286,11 +286,11 @@ data class ScanIssue(
     fun hasAIFix(): Boolean {
         return when (this.additionalData.getProductType()) {
             ProductType.OSS ->
-                return this.additionalData.isUpgradable == true
-            ProductType.CODE_SECURITY, ProductType.CODE_QUALITY -> {
-                return this.additionalData.hasAIFix
-            }
+                this.additionalData.isUpgradable
 
+            ProductType.CODE_SECURITY, ProductType.CODE_QUALITY -> {
+                this.additionalData.hasAIFix
+            }
             else -> TODO()
         }
     }
@@ -623,5 +623,6 @@ data class FolderConfigsParam(
 data class FolderConfig(
     @SerializedName("folderPath") val folderPath: String,
     @SerializedName("baseBranch") val baseBranch: String,
-    @SerializedName("localBranches") val localBranches: List<String> = emptyList()
+    @SerializedName("localBranches") val localBranches: List<String> = emptyList(),
+    @SerializedName("additionalParameters") val additionalParameters: List<String> = emptyList()
 )